deprecated_word2vec/requirements.txt
pandas
numpy
streamlit
gensim
st-star-rating
IMDbPY
tqdm